requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
executing==2.1.0
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
gradio==3.16.2
gradio_client==1.4.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.10.0
nest-asyncio==1.6.0
numpy==2.1.2
openai==0.28.0
orjson==3.10.9
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
pure_eval==0.2.3
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
ruff==0.7.0
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smartsheet-python-sdk==3.0.4
sniffio==1.3.1
stack-data==0.6.3
starlette==0.40.0
threadpoolctl==3.5.0
tomlkit==0.12.0
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.32.0
waitress==3.0.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
yarl==1.15.5
SQLAlchemy
psycopg2-binary